+++ language_server/test/test.cpp
@@ -14,9 +14,9 @@
 
 #include "gmock/gmock.h"
 
-#include "feature_language_features_test.h"
-#include "workspace_folders_test.h"
-#include "feature_text_synchronization_test.h"
+#include "lsp/feature_language_features_test.h"
+#include "lsp/workspace_folders_test.h"
+#include "lsp/feature_text_synchronization_test.h"
 #include "server_test.h"
 #include "regress_test.h"
