+++ language_server/test/ws_mngr_mock.h
@@ -17,7 +17,7 @@
 
 #include "gmock/gmock.h"
 
-#include "shared/workspace_manager.h"
+#include "workspace_manager.h"
 
 using namespace hlasm_plugin::parser_library;
